# server/docs/swagger.yaml
basePath: /
definitions:
...
  apidocs.CreateUserRequest:
    properties:
      display_name:
        example: Staff User
        type: string
      email:
        example: staff@tofex.com
        type: string
      password:
        example: changeme1234
        type: string
      role_ids:
        example:
        - 2
        items:
          type: integer
        type: array
      username:
        example: staff1
        type: string
    type: object
  apidocs.CreateUserResponse:
    properties:
      user_id:
        example: 2
        type: integer
    type: object
  apidocs.ErrorBody:
    properties:
      code:
        example: validation_error
        type: string
      message:
        example: invalid request
        type: string
    type: object
  apidocs.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apidocs.ErrorBody'
    type: object
...
  apidocs.MessageResponse:
    properties:
      message:
        example: ok
        type: string
    type: object
...
  apidocs.PatchUserRequest:
    properties:
      display_name:
        type: string
      password:
        type: string
      role_ids:
        items:
          type: integer
        type: array
    type: object
...
  apidocs.UsersListResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      label:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.RolePermission'
        type: array
      updated_at:
        type: string
    type: object
  models.RolePermission:
    properties:
      permission:
        type: string
      role_id:
        type: integer
    type: object
  models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: integer
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for Tofex restaurant/confectionery chain (staff admin + public
    guest ordering).
  title: Tofex API
  version: "1.0"
paths:
...
  /api/v1/admin/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - admin
...
  /api/v1/admin/users:
    get:
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.UsersListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      parameters:
      - description: New user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apidocs.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - admin-users
  /api/v1/admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: no content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - admin-users
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - admin-users
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apidocs.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apidocs.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidocs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - admin-users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"